Graph overhaul and wind addition
In order to switch to wind graph, I needed to overhaul how the
WeatherGraph worked. This included adding a GraphSettings class
that now handles translations, as well as all of the dimensional
information about the graph.

Also, I needed to add some buttons on the page to switch
between the modes. These don't look great right now, but they
will do for now.